+++ src/app/views/events/evidence/edit/evidenceEditBasic.js
@@ -209,7 +209,8 @@
             { value: '', label: 'Please select an Evidence Type' },
             { value: 'Predictive', label: 'Predictive' },
             { value: 'Diagnostic', label: 'Diagnostic' },
-            { value: 'Prognostic', label: 'Prognostic' }
+            { value: 'Prognostic', label: 'Prognostic' },
+            { value: 'Predisposing', label: 'Predisposing' }
           ],
           onChange: function(value, options, scope) {
             // reset clinical_significance, as its options will change
@@ -234,7 +235,8 @@
             attributeDefinitions: {
               'Predictive': 'Evidence pertains to a variant\'s effect on therapeutic response',
               'Diagnostic': 'Evidence pertains to a variant\'s impact on patient diagnosis',
-              'Prognostic': 'Evidence pertains to a variant\'s impact on disease progression, severity, or patient survival'
+              'Prognostic': 'Evidence pertains to a variant\'s impact on disease progression, severity, or patient survival',
+              'Predisposing': 'Evidence pertains to a variant\'s impact on a patient\'s likelihood to contract a disease'
             }
           }
         }
@@ -313,6 +315,10 @@
               'Prognostic': {
                 'Supports': 'The experiment or study supports a variant\'s impact on prognostic outcome',
                 'Does Not Support': 'The experiment or study does not support a prognostic association between variant and outcome'
+              },
+              'Predisposing': {
+                'Supports': 'The experiment or study supports a variant\'s likelihood to contract a disease',
+                'Does Not Support': 'The experiment or study does not support a variant\'s likelihood to contract a disease'
               }
             },
             updateDefinition: function(value, options, scope) {
@@ -351,6 +357,8 @@
             { type: 'Prognostic', value: 'Poor Outcome', label: 'Poor Outcome' },
             { type: 'Diagnostic', value: 'Positive', label: 'Positive' },
             { type: 'Diagnostic', value: 'Negative', label: 'Negative' },
+            { type: 'Predisposing', value: 'Positive', label: 'Positive' },
+            { type: 'Predisposing', value: 'Negative', label: 'Negative' },
             { type: 'N/A', value: 'N/A', label: 'N/A' }
           ],
           ngOptions: 'option["value"] as option["label"] for option in to.options',
@@ -363,6 +371,8 @@
             { type: 'Prognostic', value: 'Poor Outcome', label: 'Poor Outcome' },
             { type: 'Diagnostic', value: 'Positive', label: 'Positive' },
             { type: 'Diagnostic', value: 'Negative', label: 'Negative' },
+            { type: 'Predisposing', value: 'Positive', label: 'Positive' },
+            { type: 'Predisposing', value: 'Negative', label: 'Negative' },
             { type: 'N/A', value: 'N/A', label: 'N/A' }
           ],
           helpText: 'Positive or negative association of the Variant with predictive, prognostic, or diagnostic evidence types. If the variant was not associated with a positive or negative outcome, N/A should be selected. Evidence Type must be selected before this field is enabled.',
@@ -452,7 +462,7 @@
           ],
           valueProp: 'value',
           labelProp: 'label',
-          helpText: 'If multiple drugs were added, please select an interaction type that describes their relationship to each other.',
+          helpText: 'Please indicate whether the drugs specified above are substitutes, or are used in sequential or combination treatments.',
           data: {
             attributeDefinition: '&nbsp;',
             attributeDefinitions: {