+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 * variants now show aliases and HGVS expressions
 * variant edit form provides multi-inputs for HGVS expressions and aliases
 * gene advanced search now provides Suggested Revisions attribute
+* Evidence Type fields on add/edit evidence item forms now provide a Predisposing type
 * gene grid variant count column now filterable and sortable
 * meetings page updates for NKI
 * added custom header to client requests so server can better differentiate between client and direct API access
+* variant name on add evidence for now a required field
 
 0.0.5 / 2016-08-26
 ==================